Introduce resilience to rpc client and lock implementation
The network may re-order, delay or drop our rpc requests. Before this
commit, our client did not have a mechanism to recover from this. We may
erroneously let multiple clients acquire a lock or not release one.

this commit introduces the ability to retry infinitely. it does not have a
backoff but our test harness takes care of not getting stuck in an infinite loop.

we simply retry a get or a put as many times as possible until we get a true response
from the server

we also modify the lock to work in the scenario where the first put call succeeded but
the reply rpc was dropped by the network. in that case, we should not let a client acquire
a lock